requirements.txt
anthropic>=0.39.0
openai>=1.54.0
google-generativeai>=0.8.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.149.0
chromadb>=0.5.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
tiktoken>=0.7.0
numpy>=1.24.0
